pyproject.toml
[project]
name = "pgtrigger-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "django>=5.2",
    "django-pgtrigger>=4.15.0",
    "psycopg[binary]>=3.2.6",
]
